(refactor) moved the view details away from patient chart
