+++ packages/esm-morgue-app/src/view-details/main/main.scss
