+++ packages/esm-morgue-app/src/view-details/panels/attachement.component.tsx
