+++ packages/esm-morgue-app/src/view-details/panels/observations/observation.component.tsx
