+++ packages/esm-morgue-app/src/view-details/panels/panels.scss
