+++ packages/esm-morgue-app/src/view-details/views-details.scss
