+++ packages/esm-morgue-app/translations/en.json
@@ -1,17 +1,13 @@
 {
   "action": "Action",
-  "address": "Address",
   "admissionDate": "Admission Date",
   "admissionForm": "Admission form",
   "admit": "Admit",
-  "admitBodies": "Admit bodies",
   "admitBody": "Admit",
   "admitted": "Admitted",
   "admittedHeader": "Admitted",
   "age": "Age",
   "allocation": "Allocation",
-  "allocations": "Allocation",
-  "assignBed": "Assign Bed",
   "available": "Available",
   "awaiting": "Awaiting",
   "awaitingAdmissionHeader": "Awaiting Admission",
@@ -23,59 +19,59 @@
   "compartment": "Compartment",
   "compartmentNumber": "Compartment Number",
   "compartmentShare": "Compartment Share",
+  "compartmentSwap": "Compartment swap",
   "dateAndTimeofDeath": "Date and time of death",
   "dateOfAdmission": "Date of admission",
+  "dateOfDeath": "Date of death",
   "dateQueued": "Date queued",
   "day": "Day",
   "days": "Days",
   "daysInMortuary": "Days in mortuary",
-  "deceased": "Deceased",
-  "deceasedAlreadyAdmitted": "Deceased patient Already Admitted",
-  "deceasedHasAlreadyBeenAdmitted": "Deceased has already been admitted",
+  "daysSinceDeath": "Days Since Death",
   "discharge": "Discharge body",
-  "dischargeBodies": "Discharged bodies",
-  "dischargeCard": "Discharge - Card View",
   "discharged": "Discharged",
   "dischargedHeader": "Discharged",
   "dischargeForm": "Discharge form",
-  "dischargeList": "Discharge - List View",
+  "dispose": "Dispose",
+  "disposeForm": "Dispose",
   "dividerText": "Shared",
   "durationOnWard": "Days In Queue",
-  "edit": "Edit",
   "empty": "Empty",
   "emptyCompartment": "Empty compartment",
-  "errorLoadingData": "Error loading data",
+  "errorLoadingPatients": "Error loading discharged patients",
   "errorMessage": "Error",
   "gender": "Gender",
   "idNumber": "ID Number",
   "lengthofStay": "Length of stay",
   "listView": "List",
-  "loading": "Loading...",
   "loadingPatients": "Loading patients...",
   "mortuary": "Mortuary",
   "mortuaryOperations": "Mortuary operations",
   "name": "Name",
-  "nextOfKin": "Next of Kin",
-  "noAdmittedBodies": "There are no admitted bodies matching your search",
-  "noAvailable": "No available",
+  "noAdmittedPatients": "No admitted patients found",
+  "noAdmittedPatientsDescription": "There are currently no admitted patients to display.",
+  "noAwaitingPatients": "No awaiting patients found",
+  "noAwaitingPatientsDescription": "There are currently no awaiting patients to display.",
   "noDeceasedPatients": "No deceased patients awaiting admission",
-  "noDeceasedPersons": "There are no deceased persons on the waiting list",
-  "noResultFound": "Sorry, no results found",
-  "noWaitingList": "Waiting List",
+  "noDischargedPatients": "No discharged patients found",
+  "noDischargedPatientsDescription": "There are currently no discharged patients to display.",
   "occupied": "Occupied",
-  "patientAlreadyAdmittedMessage": "This deceased patient has already been admitted.",
+  "patientDetails": "Patient Details",
   "patientName": "Patient Name",
-  "phone": "Phone",
-  "searchForADeceased": "Try to search again using the deceased patient's unique ID number",
+  "postmortem": "Postmortem",
+  "postmortemForm": "Postmortem",
+  "printGatePass": "Gate Pass",
+  "printPostmortem": "Postmortem",
+  "printPostmortemReport": "Postmortem Report",
   "searchPatients": "Search Patients",
   "sharedCompartment_one": "{{count}} sharing",
   "sharedCompartment_other": "{{count}} sharing",
-  "sharing": "Sharing",
   "singleOccupancy": "Single",
   "status": "Status",
   "swapCompartment": "Swap compartment",
   "timeSpent": "Time spent",
   "totalCount": "total",
   "unknown": "Unknown",
+  "viewDetails": "View details",
   "yearsOld": "Yrs old"
 }
